#3 feat: UI와 ItemManager 데이터 연결
