+++ OpenMarket/OpenMarket/Enum/ViewMode.swift
@@ -1,0 +1,5 @@
+import Foundation
+
+enum ViewMode {
+    case list, grid
+}
