+++ OpenMarket/OpenMarket/Protocols/ViewControllerDelegate.swift
@@ -1,0 +1,6 @@
+import Foundation
+
+protocol ViewControllerDelegate {
+    func didTapViewMode()
+    func reloadCollectionView()
+}
